+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HubConnection.java
@@ -16,11 +16,6 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Consumer;
 
-import okhttp3.Cookie;
-import okhttp3.CookieJar;
-import okhttp3.HttpUrl;
-import okhttp3.OkHttpClient;
-
 public class HubConnection {
     private String url;
     private Transport transport;
@@ -57,7 +52,7 @@ public class HubConnection {
         }
         this.httpClient = client;
         if (client == null) {
-            this.httpClient = new DefaultHttpClient();
+            this.httpClient = new DefaultHttpClient(this.logger);
         }
 
         if (transport != null) {
@@ -65,58 +60,6 @@ public class HubConnection {
         }
 
         this.skipNegotiate = skipNegotiate;
-
-        this.httpClient = new OkHttpClient.Builder()
-                .cookieJar(new CookieJar() {
-                    private List<Cookie> cookieList = new ArrayList<>();
-                    private Lock cookieLock = new ReentrantLock();
-
-                    @Override
-                    public void saveFromResponse(HttpUrl url, List<Cookie> cookies) {
-                        cookieLock.lock();
-                        try {
-                            for (Cookie cookie : cookies) {
-                                boolean replacedCookie = false;
-                                for (int i = 0; i < cookieList.size(); i++) {
-                                    Cookie innerCookie = cookieList.get(i);
-                                    if (cookie.name().equals(innerCookie.name()) && innerCookie.matches(url)) {
-                                        // We have a new cookie that matches an older one so we replace the older one.
-                                        cookieList.set(i, innerCookie);
-                                        replacedCookie = true;
-                                        break;
-                                    }
-                                }
-                                if (!replacedCookie) {
-                                    cookieList.add(cookie);
-                                }
-                            }
-                        } finally {
-                            cookieLock.unlock();
-                        }
-                    }
-
-                    @Override
-                    public List<Cookie> loadForRequest(HttpUrl url) {
-                        cookieLock.lock();
-                        try {
-                            List<Cookie> matchedCookies = new ArrayList<>();
-                            List<Cookie> expiredCookies = new ArrayList<>();
-                            for (Cookie cookie : cookieList) {
-                                if (cookie.expiresAt() < System.currentTimeMillis()) {
-                                    expiredCookies.add(cookie);
-                                } else if (cookie.matches(url)) {
-                                    matchedCookies.add(cookie);
-                                }
-                            }
-
-                            cookieList.removeAll(expiredCookies);
-                            return matchedCookies;
-                        } finally {
-                            cookieLock.unlock();
-                        }
-                    }
-                })
-                .build();
 
         this.callback = (payload) -> {
